remove:(atomic) smarter...

+++ include/interrupter.h
@@ -54,8 +54,8 @@ struct nco r_osc = {.phase_accum = 0,
 const uint_fast32_t dac1_clock = 198'000;
 // const uint_fast32_t dac1_clock = 44'000;
 const uint_fast32_t adc1_clock = 44'000;
-unsigned int dac_double_buff[256] = {0};
-unsigned int dac_double_buff2[256] = {0};
+uint16_t dac_double_buff[256] = {0};
+uint16_t dac_double_buff2[256] = {0};
 
 struct timer tim6_settings = {.timx=TIM6, .apb_clock_reg=LL_APB1_GRP1_PERIPH_TIM6, .trigger_output=LL_TIM_TRGO_UPDATE};
 struct timer tim7_settings = {.timx=TIM7, .apb_clock_reg=LL_APB1_GRP1_PERIPH_TIM7, .trigger_output=LL_TIM_TRGO_UPDATE};
