+++ include/nco.h
@@ -8,7 +8,7 @@ enum cyrcles {entrance, first, second, third, fourth, fifth, sixth, seventh,
 
 struct phase_distortion{
     volatile bool on;
-    volatile unsigned int amount;
+    volatile uint16_t amount;
     enum cyrcles dante;
     enum cyrcles past_dante;
     volatile uint32_t distortion_value;
@@ -35,16 +35,16 @@ void generate_half_signal(volatile const uint16_t data[static 128],
                           uint16_t sectionLength, struct nco nco[static 1]);
 
 void update_data_buff (const uint16_t data[static 128],
-                            unsigned int bufferSection[static 128],
+                            uint16_t bufferSection[static 128],
                             uint16_t sectionLength);
 #pragma GCC diagnostic ignored "-Wignored-qualifiers"
 
-__attribute__((pure)) unsigned int map_12b_to_distortion_amount(uint16_t value);
-__attribute__((pure))unsigned int map_12b_to_hz(uint16_t adc_value, enum freq_modes pitch_modes);
+__attribute__((pure)) uint16_t map_12b_to_distortion_amount(uint16_t value);
+__attribute__((pure))uint16_t map_12b_to_hz(uint16_t adc_value, enum freq_modes pitch_modes);
 bool stage_pending_inc(volatile uint16_t adc_raw_value, struct nco nco[static 1], const uint_fast32_t sample_rate);
-void stage_modulated_signal_values(struct nco osc[static 1], unsigned int distortion_cv, volatile uint16_t pitch_cv, uint32_t master_clock);
+void stage_modulated_signal_values(struct nco osc[static 1], uint16_t distortion_cv, volatile uint16_t pitch_cv, uint32_t master_clock);
 
-__attribute__((pure, always_inline)) inline static uint64_t compute_nco_increment(unsigned int note, const uint_fast32_t sample_rate){
+__attribute__((pure, always_inline)) inline static uint64_t compute_nco_increment(uint16_t note, const uint_fast32_t sample_rate){
     int32_t tmp = ((note * (1<<16))/sample_rate);
     return (tmp<<16);
 }
